Refactored dram pmem execution.

+++ src/parallel_benchmark.cpp
@@ -39,10 +39,12 @@ void ParallelBenchmark::tear_down(bool force) {
   for (size_t index = 0; index < pmem_data_.size(); index++) {
     if (pmem_data_[index] != nullptr) {
       munmap(pmem_data_[index], configs_[index].total_memory_range);
-      close(file_descriptors_[index]);
       pmem_data_[index] = nullptr;
+      if (configs_[index].is_pmem) {
+        close(file_descriptors_[index]);
+      }
     }
-    if (owns_pmem_files_[index] || force) {
+    if (configs_[index].is_pmem && (owns_pmem_files_[index] || force)) {
       std::filesystem::remove(pmem_files_[index]);
     }
   }
